+++ Pandora/Paradigm/Inventory/Accumulator.hs
@@ -2,7 +2,6 @@
 
 module Pandora.Paradigm.Inventory.Accumulator (Accumulator (..), Accumulated, gather) where
 
-import Pandora.Core.Functor (Variant (Co))
 import Pandora.Paradigm.Basis.Product (Product ((:*:)), type (:*:))
 import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, unwrap))
 import Pandora.Paradigm.Controlflow.Joint.Transformer.Monadic (Monadic (lay, wrap), (:>) (TM))
@@ -35,7 +34,7 @@ instance Semigroup e => Bindable (Accumulator e) where
 	Accumulator (e :*: x) >>= f = let (e' :*: b) = unwrap $ f x in
 		Accumulator $ e + e':*: b
 
-type instance Schematic Monad (Accumulator e) u = UT 'Co 'Co ((:*:) e) u
+type instance Schematic Monad (Accumulator e) u = UT Covariant Covariant ((:*:) e) u
 
 instance Interpreted (Accumulator e) where
 	type Primary (Accumulator e) a = e :*: a
@@ -47,17 +46,17 @@ instance Monoid e => Monadic (Accumulator e) where
 
 type Accumulated e t = Adaptable (Accumulator e) t
 
-instance Covariant u => Covariant (UT 'Co 'Co ((:*:) e) u) where
+instance Covariant u => Covariant (UT Covariant Covariant ((:*:) e) u) where
 	f <$> UT x = UT $ f <$$> x
 
-instance (Semigroup e, Applicative u) => Applicative (UT 'Co 'Co ((:*:) e) u) where
+instance (Semigroup e, Applicative u) => Applicative (UT Covariant Covariant ((:*:) e) u) where
 	UT f <*> UT x = UT $ k <$> f <*> x where
 		k ~(u :*: g) ~(v :*: y) = u + v :*: g y
 
-instance (Pointable u, Monoid e) => Pointable (UT 'Co 'Co ((:*:) e) u) where
+instance (Pointable u, Monoid e) => Pointable (UT Covariant Covariant ((:*:) e) u) where
 	point = UT . point . (zero :*:)
 
-instance (Semigroup e, Pointable u, Bindable u) => Bindable (UT 'Co 'Co ((:*:) e) u) where
+instance (Semigroup e, Pointable u, Bindable u) => Bindable (UT Covariant Covariant ((:*:) e) u) where
 	UT x >>= f = UT $ x >>= \(acc :*: v) -> (\(acc' :*: y) -> (acc + acc' :*: y)) <$> unwrap (f v)
 
 gather :: (Covariant t, Accumulated e t) => e -> t ()