+++ CHANGELOG.md
@@ -149,4 +149,5 @@
 * Define `Comonadic` type class for comonad transformers
 * Define `Transformer` umbrella module and type family
 * Rename `:<` data constructor to `Twister` to not confuse it with comonad transformer type operator
-* Use `UT` joint scheme for `Stack` data structure
+* Use `UT` joint scheme for `Stack` and `Graph` data structures
+* Remove `Variant` type in favor of `Covariant` and `Contravariant` constraints in joint schemes